See ...\DOCUMENTATION\RefactoringDocumentation.docx
Added some classes to ease testing and removed all methods from the main
class to the new ones.
Some methods need to be extracted from Main() as it is way too big.
Added a unit testing project.
Added a docx file for documenting all changes that we make (as we make
them so that we don't forget later).
Added the original project we started with (in the task description it
says we need to give that along with our project too).