Minor changes
Added back a check in ScoreBoard AddScore(). Without it the game won't
record high scores if there are less than the max entries currently and
the latest one is lower than the last recorder (which could be the first
one). Example - if with the first game a high score of 10 is set, the
following scores won't be recorded if they are less than that.