Added some interfaces and fixed saving problem
Added IGameField and IScoreBoard interfaces and code to use them
everywhere the classes were user.
Fixed a problem with saving where after restore the game could not get
finished because it only ended when the score becomes 35 which is not
possible after restore.
Did all the changes StyleCop pointed out and changed properties to look
consistent throughout the code.